move team views, update/fix tests
